# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API'S",
    "description": "My API Microservice Project documentation",
    "version": "1.0.0"
  },
  "host": "0.0.0.0:3000",
  "basePath": "",
  "tags": [
    {
      "name": "Users",
      "description": "Operations related to users"
    },
    {
      "name": "Games",
      "description": "Operations related to games"
    },
    {
      "name": "Orders",
      "description": "Operations related to orders"
    },
    {
      "name": "Authentication",
      "description": "Operations related to authentication"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieve all users",
        "description": "Retrieve a list of all users",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieve one user by id",
        "description": "Retrieve a user by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user",
        "description": "Update an existing user",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "user",
            "description": "Updated user object",
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user",
        "description": "Delete an existing user",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/games": {
      "get": {
        "tags": [
          "Games"
        ],
        "summary": "Retrieve all games",
        "description": "Retrieve a list of all games",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Games"
        ],
        "summary": "Add a new game",
        "description": "Adds a new game to the system",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "game",
            "description": "Game object",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          }
        ]
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Create a new user",
        "description": "Create a new user to the system",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ]
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User login",
        "description": "Allows a user to log in to the system",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "credentials",
            "description": "User credentials",
            "schema": {
              "$ref": "#/definitions/UserCredentials"
            }
          }
        ]
      }
    },
    "/games/{id}": {
      "get": {
        "tags": [
          "Games"
        ],
        "summary": "Retrieve one game by id",
        "description": "Retrieve a game by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Retrieve all orders",
        "description": "Retrieve a list of all orders",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Add a new order",
        "description": "Adds a new order to the system",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "order",
            "description": "Order object",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Retrieve one order by id",
        "description": "Retrieve an order by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "UserCredentials": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "nom": {
          "type": "string"
        },
        "prenom": {
          "type": "string"
        },
        "age": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UpdateUser": {
      "type": "object",
      "properties": {
        "nom": {
          "type": "string"
        },
        "prenom": {
          "type": "string"
        },
        "age": {
          "type": "string"
        }
      }
    },
    "Game": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "prix": {
          "type": "string"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "idUser": {
          "type": "string"
        },
        "idGames": {
          "type": "string"
        },
        "total": {
          "type": "string"
        }
      }
    }
  }
}